+++ packages/automated_actions_utils/automated_actions_utils/gql_definitions/fragments/vault_secret.gql
