+++ packages/automated_actions_utils/tests/test_cluster_connection.py
@@ -1,38 +1,27 @@
-# ruff: noqa: S105, S106
+# ruff: noqa:  S106
 from unittest.mock import MagicMock
 
 import pytest
+from automated_actions.config import settings
 from pytest_mock import MockerFixture
 
-from automated_actions.utils.cluster_connection import (
+from automated_actions_utils.cluster_connection import (
     ClusterConnectionData,
     ClusterMissingInAppInterfaceError,
     MissingAppInterfaceClusterAutomationTokenError,
     get_cluster_connection_data,
 )
-from automated_actions.utils.vault_client import SecretFieldNotFoundError
+from automated_actions_utils.vault_client import SecretFieldNotFoundError
 
 
 @pytest.fixture(autouse=True)
 def mock_gql_client(mocker: MockerFixture) -> MagicMock:
-    return mocker.patch("automated_actions.utils.cluster_connection.GQLClient")
+    return mocker.patch("automated_actions_utils.cluster_connection.GQLClient")
 
 
 @pytest.fixture(autouse=True)
 def mock_vault_client(mocker: MockerFixture) -> MagicMock:
-    return mocker.patch("automated_actions.utils.cluster_connection.VaultClient")
-
-
-@pytest.fixture(autouse=True)
-def mock_settings(mocker: MockerFixture) -> None:
-    mock = mocker.patch("automated_actions.utils.cluster_connection.settings")
-    mock.qontract_server_url = "https://qontract.example.com"
-    mock.qontract_server_token = "qontract-token"
-    mock.vault_server_url = "https://vault.example.com"
-    mock.vault_role_id = "role-id"
-    mock.vault_secret_id = "secret-id"
-    mock.vault_kube_auth_role = None
-    mock.vault_kube_auth_mount = None
+    return mocker.patch("automated_actions_utils.cluster_connection.VaultClient")
 
 
 def test_get_cluster_connection_data_success(
@@ -57,7 +46,7 @@ def test_get_cluster_connection_data_success(
     # Mock Vault client response
     mock_vault_client.return_value.read_secret.return_value = {"token": "test-token"}
 
-    result = get_cluster_connection_data("test-cluster")
+    result = get_cluster_connection_data("test-cluster", settings)
 
     assert result == ClusterConnectionData(
         url="https://cluster.example.com", token="test-token"
@@ -73,7 +62,7 @@ def test_get_cluster_connection_data_missing_cluster(
     mock_gql_client.return_value.query.return_value = {"cluster": []}
 
     with pytest.raises(ClusterMissingInAppInterfaceError) as exc_info:
-        get_cluster_connection_data("missing-cluster")
+        get_cluster_connection_data("missing-cluster", settings)
 
     assert str(exc_info.value) == "cluster 'missing-cluster' missing in app-interface"
     mock_gql_client.return_value.query.assert_called_once()
@@ -94,7 +83,7 @@ def test_get_cluster_connection_data_missing_automation_token(
     }
 
     with pytest.raises(MissingAppInterfaceClusterAutomationTokenError) as exc_info:
-        get_cluster_connection_data("test-cluster")
+        get_cluster_connection_data("test-cluster", settings)
 
     assert str(exc_info.value) == "No automationToken for cluster test-cluster"
     mock_gql_client.return_value.query.assert_called_once()
@@ -123,7 +112,7 @@ def test_get_cluster_connection_data_secret_field_not_found(
     mock_vault_client.return_value.read_secret.return_value = {}
 
     with pytest.raises(SecretFieldNotFoundError) as exc_info:
-        get_cluster_connection_data("test-cluster")
+        get_cluster_connection_data("test-cluster", settings)
 
     assert str(exc_info.value) == ("token not found in secret secret/path")
     mock_gql_client.return_value.query.assert_called_once()
